+++ src/services/integrations/stripe.js
@@ -1,5 +1,7 @@
 const stripe = require('stripe')(process.env.STRIPE_API_SECRET_KEY, stripeOptions)
 
+const webhookSignatureHeader = 'stripe-signature'
+
 async function createCheckoutSession(lineItems) {
   const payload = checkoutSessionCreationPayload(lineItems)
   return await stripe.checkout.sessions.create(payload)
@@ -11,6 +13,14 @@ async function checkoutSession(sessionId) {
 
 async function promotionCode(promotionId) {
   return await stripe.promotionCodes.retrieve(promotionId)
+}
+
+function constructWebhookEvent(payload, payloadHeaders, secret) {
+  return stripe.webhooks.constructEvent(payload, payloadHeaders[webhookSignatureHeader], secret)
+}
+
+function webhookEventData(event) {
+  return event.data.object
 }
 
 function checkoutSessionCreationPayload(lineItems) {
@@ -46,5 +56,7 @@ function stripeOptions() {
 module.exports = {
   createCheckoutSession,
   checkoutSession,
-  promotionCode
+  promotionCode,
+  constructWebhookEvent,
+  webhookEventData
 }